+++ src/main/java/tornadofx/Properties.kt
@@ -57,14 +57,18 @@ fun Class<*>.findMethodByName(name: String): Method? {
 /**
  * Convert an owner instance and a corresponding property reference into an observable
  */
-fun <S, T> S.observable(prop: KMutableProperty1<S, T>): ObjectProperty<T> {
-    val owner = this
+fun <S, T> S.observable(prop: KMutableProperty1<S, T>) = observable(this, prop)
+
+/**
+ * Convert an owner instance and a corresponding property reference into an observable
+ */
+@JvmName("observableFromMutableProperty")
+fun <S, T> observable(owner: S, prop: KMutableProperty1<S, T>): ObjectProperty<T> {
     return object : SimpleObjectProperty<T>(owner, prop.name) {
         override fun get() = prop.get(owner)
         override fun set(v: T) = prop.set(owner, v)
     }
 }
-
 /**
  * Convert an owner instance and a corresponding property reference into a readonly observable
  */